# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: 찾아줄게!
  description: 6팀 데이터 분석 프로젝트 API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /items/{id}:   
    patch:
      tags:
        - Item
      summary: 분실물 수정
      operationId: editItem
      parameters:
        - name: id
          in: path
          description: Patch 방식을 통해 특정 분실물 수정
          required: true
          example: 17
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                data:
                  type: object
                  properties:
                    name:
                      type: string
                      example: 반스 숄더백
                    description:
                      type: string
                      example: 카드가 들어있는 반스 숄더백을 보관하고 있습니다.
                    status:
                      type: string
                      example: pickup
                    area:
                      type: string
                      example: 도봉구
                    takePlace:
                      type: string
                      example: 회사 내 분실 센터  
                    lostPlace:
                      type: string
                      example: 경일운수
                    getDate:
                      type: string
                      example: "2023-03-16"
                    categoryName:
                      type: string
                      example: 핸드폰
                    image:
                      type: string
                      format: binary
      responses:
        "200":
          description: 분실물 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 17
                      name:
                        type: string
                        example: 반스 숄더백
                      description:
                        type: string
                        example: 카드가 들어있는 반스 숄더백을 보관하고 있습니다.
                      status:
                        type: string
                        example: pickup
                      categoryId:
                        type: integer
                        example: 3
                      editedUserId:
                        type: integer
                        example: 3
                      pickupUserId:
                        type: integer
                        example: 1
                      imageId:
                        type: integer
                        example: 4
                      pickupDate:
                        type: string
                        format: date-time
                        example: "2023-03-16T15:53:36.345Z"
                      getDate:
                        type: string
                        format: date-time
                        example: "2023-03-16T00:00:00"
                      area:
                        type: string
                        example: 도봉구
                      takePlace:
                        type: string
                        example: 회사 내 분실 센터  
                      lostPlace:
                        type: string
                        example: 경일운수
                      createdAt:
                        type: string
                        format: date-time
                        example: "2023-03-16T15:07:16.218Z"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "2023-03-16T15:48:37.765Z"
        "400":
          description: 분실물 수정 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode:
                    type: string
                    default: ITEM_EDIT_ERROR
                  errorMessage:
                    type: string
                    default: 분실물 수정에 실패했습니다.  
    get:
      tags:
        - Board
      summary: 찾아주세요 게시글 상세조회
      parameters:
        - name: id
          in: path
          description: 특정 게시글 id
          required: true
          example: 17
          schema:
            type: string
      responses:
        "200":
          description: 상세조회 성공
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        default: 2
                      name:
                        type: string
                        default: 휴대폰 찾아주세요!
                      description:
                        type: string
                        default: 36개월 할부 휴대폰이 사라졌어요..
                      status:
                        type: string
                        default: store
                      lostPlace:
                        type: string
                        default: 집앞
                      takePlace:
                        type: string
                        default: 분실물센터
                      categoryName:
                        type: string
                        default: 핸드폰
                      userName:
                        type: string
                        default: 달달
                      imageUrl:
                        type: string
                        default: uploads/medium/1679294981674.jpg  
                      getDate:
                        type: string
                        default: "2023-03-19"
                      createdAt:
                        type: string
                        default: "2023-03-20"
        "400":
          description: 분실물 상세조회 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode:
                    type: string
                    default: ITEM_FIND_ERROR
                  errorMessage:
                    type: string
                    default: 분실물 상세 조회에 실패했습니다.  
    delete:
      tags:
        - Item
      summary: 분실물 삭제
      operationId: deleteItem
      parameters:
        - name: id
          in: path
          description: Delete 방식을 통해 특정 분실물 삭제
          required: true
          example: 17
          schema:
            type: string        
      responses:
        "200":
          description: 분실물 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    default: 분실물 삭제가 완료되었습니다.
        "400":
          description: 분실물 삭제 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorCode:
                    type: string
                    default: ITEM_DELETE_ERROR
                  errorMessage:
                    type: string
                    default: 분실물 삭제에 실패했습니다. 
...